// .dist/planning-v2-desktop/src/lib/planning/core/v2/types.ts
export type ValidationIssueV2 = {
  path: string;
  message: string;
};

export class PlanningV2ValidationError extends Error {
  readonly issues: ValidationIssueV2[];

  constructor(message: string, issues: ValidationIssueV2[]) {
    super(message);
    this.name = "PlanningV2ValidationError";
    this.issues = issues;
  }
}